requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 26.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
a2wsgi=1.10.10=pypi_0
aiosmtplib=5.1.0=pypi_0
aiosqlite=0.21.0=pypi_0
alembic=1.18.4=pypi_0
amqp=5.3.1=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.12.1=pypi_0
apache-airflow=3.1.8=pypi_0
apache-airflow-core=3.1.8=pypi_0
apache-airflow-providers-celery=3.17.0=pypi_0
apache-airflow-providers-common-compat=1.14.0=pypi_0
apache-airflow-providers-common-io=1.7.1=pypi_0
apache-airflow-providers-common-sql=1.32.0=pypi_0
apache-airflow-providers-smtp=2.4.2=pypi_0
apache-airflow-providers-standard=1.12.0=pypi_0
apache-airflow-task-sdk=1.1.8=pypi_0
argcomplete=3.6.3=pypi_0
argon2-cffi=25.1.0=pypi_0
argon2-cffi-bindings=25.1.0=pypi_0
arrow=1.4.0=pypi_0
asgiref=3.11.1=pypi_0
asttokens=3.0.1=pypi_0
async-lru=2.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=25.4.0=pypi_0
babel=2.18.0=pypi_0
backports-strenum=1.3.1=pypi_0
beautifulsoup4=4.14.3=pypi_0
billiard=4.2.4=pypi_0
bleach=6.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.12.2=h06a4308_0
cadwyn=5.4.6=pypi_0
celery=5.6.2=pypi_0
certifi=2026.2.25=pypi_0
cffi=2.0.0=pypi_0
charset-normalizer=3.4.5=pypi_0
click=8.3.1=pypi_0
click-didyoumean=0.3.1=pypi_0
click-plugins=1.1.1.2=pypi_0
click-repl=0.3.0=pypi_0
colorlog=6.10.1=pypi_0
comm=0.2.3=pypi_0
cron-descriptor=2.0.6=pypi_0
croniter=6.0.0=pypi_0
cryptography=42.0.8=pypi_0
debugpy=1.8.20=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.3.1=pypi_0
dill=0.4.1=pypi_0
dnspython=2.8.0=pypi_0
email-validator=2.3.0=pypi_0
exceptiongroup=1.3.1=pypi_0
executing=2.2.1=pypi_0
fastapi=0.117.1=pypi_0
fastapi-cli=0.0.24=pypi_0
fastjsonschema=2.21.2=pypi_0
flower=2.0.1=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2026.2.0=pypi_0
google-api-core=2.30.0=pypi_0
google-auth=2.49.1=pypi_0
google-cloud-core=2.5.0=pypi_0
google-cloud-storage=3.10.1=pypi_0
google-cloud-storage-control=1.11.0=pypi_0
google-crc32c=1.8.0=pypi_0
google-resumable-media=2.8.0=pypi_0
googleapis-common-protos=1.73.0=pypi_0
greenback=1.3.0=pypi_0
greenlet=3.3.2=pypi_0
grpc-google-iam-v1=0.14.4=pypi_0
grpcio=1.80.0=pypi_0
grpcio-status=1.80.0=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.7.1=pypi_0
httpx=0.28.1=pypi_0
humanize=4.15.0=pypi_0
idna=3.11=pypi_0
importlib-metadata=8.4.0=pypi_0
ipykernel=7.2.0=pypi_0
ipython=8.39.0=pypi_0
isoduration=20.11.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
json5=0.14.0=pypi_0
jsonpointer=3.1.1=pypi_0
jsonschema=4.26.0=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter-client=8.8.0=pypi_0
jupyter-core=5.9.1=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.3.1=pypi_0
jupyter-server=2.17.0=pypi_0
jupyter-server-terminals=0.5.4=pypi_0
jupyterlab=4.5.6=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.28.0=pypi_0
kombu=5.6.2=pypi_0
lark=1.3.1=pypi_0
lazy-object-proxy=1.12.0=pypi_0
ld_impl_linux-64=2.44=h9e0c5a2_3
libcst=1.8.6=pypi_0
libffi=3.3=he6710b0_2
libgcc=15.2.0=h69a1729_7
libgcc-ng=15.2.0=h166f726_7
libgomp=15.2.0=h4751f2c_7
libstdcxx=15.2.0=h39759b7_7
libstdcxx-ng=15.2.0=hc03a8fd_7
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
linkify-it-py=2.1.0=pypi_0
lockfile=0.12.2=pypi_0
mako=1.3.10=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
matplotlib-inline=0.2.1=pypi_0
mdurl=0.1.2=pypi_0
methodtools=0.4.7=pypi_0
mistune=3.2.0=pypi_0
more-itertools=10.8.0=pypi_0
msgspec=0.20.0=pypi_0
natsort=8.4.0=pypi_0
nbclient=0.10.4=pypi_0
nbconvert=7.17.0=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h7934f7d_0
nest-asyncio=1.6.0=pypi_0
notebook=7.5.5=pypi_0
notebook-shim=0.2.4=pypi_0
openssl=1.1.1w=h7f8727e_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-exporter-otlp=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.27.0=pypi_0
opentelemetry-proto=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
outcome=1.3.0.post0=pypi_0
overrides=7.7.0=pypi_0
packaging=26.0=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.6=pypi_0
pathspec=1.0.4=pypi_0
pendulum=3.2.0=pypi_0
pexpect=4.9.0=pypi_0
pip=26.0.1=pyhc872135_0
platformdirs=4.9.4=pypi_0
pluggy=1.6.0=pypi_0
prometheus-client=0.24.1=pypi_0
prompt-toolkit=3.0.52=pypi_0
proto-plus=1.27.2=pypi_0
protobuf=6.33.6=pypi_0
psutil=7.2.2=pypi_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py4j=0.10.9.9=pypi_0
pyasn1=0.6.3=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycparser=3.0=pypi_0
pydantic=2.12.5=pypi_0
pydantic-core=2.41.5=pypi_0
pygments=2.19.2=pypi_0
pygtrie=2.5.0=pypi_0
pyjwt=2.11.0=pypi_0
pyspark=4.1.1=pypi_0
python=3.10.0=h12debd9_5
python-daemon=3.1.2=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.2.2=pypi_0
python-json-logger=4.1.0=pypi_0
python-multipart=0.0.22=pypi_0
python-slugify=8.0.4=pypi_0
pytz=2026.1.post1=pypi_0
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=pypi_0
readline=8.3=hc2a1206_0
redis=6.4.0=pypi_0
referencing=0.37.0=pypi_0
requests=2.32.5=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rfc3987-syntax=1.1.0=pypi_0
rich=13.9.4=pypi_0
rich-argparse=1.7.2=pypi_0
rich-toolkit=0.19.7=pypi_0
rpds-py=0.30.0=pypi_0
send2trash=2.1.0=pypi_0
setproctitle=1.3.7=pypi_0
setuptools=80.10.2=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.8.3=pypi_0
sqlalchemy=2.0.48=pypi_0
sqlalchemy-jsonfield=1.0.2=pypi_0
sqlalchemy-utils=0.42.1=pypi_0
sqlite=3.51.2=h3e8d24a_0
sqlparse=0.5.5=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.48.0=pypi_0
structlog=25.5.0=pypi_0
svcs=25.1.0=pypi_0
tabulate=0.10.0=pypi_0
tenacity=9.1.4=pypi_0
termcolor=3.3.0=pypi_0
terminado=0.18.1=pypi_0
text-unidecode=1.3=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.15=h54e0aa7_0
tomli=2.4.0=pypi_0
tornado=6.5.4=pypi_0
traitlets=5.14.3=pypi_0
typer=0.24.1=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2025.3=pypi_0
tzlocal=5.3.1=pypi_0
uc-micro-py=2.0.0=pypi_0
universal-pathlib=0.2.6=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.6.3=pypi_0
uuid6=2025.0.1=pypi_0
uvicorn=0.41.0=pypi_0
uvloop=0.22.1=pypi_0
vine=5.1.0=pypi_0
watchfiles=1.1.1=pypi_0
wcwidth=0.6.0=pypi_0
webcolors=25.10.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.9.0=pypi_0
websockets=16.0=pypi_0
wheel=0.46.3=py310h06a4308_0
wirerope=1.0.0=pypi_0
wrapt=2.1.2=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.8.2=h448239c_0
zipp=3.23.0=pypi_0
zlib=1.3.1=hb25bd0a_0
